Send Chat Message as Bot when an outgoing raid happened
This is done by  opening a new EventSub connection for the channel.raid subscription type and using the Firebot API to send a chat message as the bot.
To register the EventSub subscription the access token from Firebot is being used together with it's reverse engineered Client-ID.

This subscription topic does not require authentication, so I could also just have used an App Access Token if I wanted to.